+++ init.el
@@ -115,6 +115,7 @@ by Prelude.")
 (require 'prelude-mode)
 (require 'prelude-editor)
 (require 'prelude-global-keybindings)
+(require 'prelude-welcome)
 
 ;; macOS specific settings
 (when (eq system-type 'darwin)